tab css; created dm file; tab now works with channel page & dm page
