+++ front/src/styles/chat.scss
@@ -6,16 +6,17 @@
 	--highlight: #8860D0;
 }
 
+$tab-color: rgb(121, 121, 121);
 
 .chat-tab {
 	font-size: 20px;
-	background-color: rgb(121, 121, 121);
+	background-color: $tab-color;
 	color: white;
 	font-family: 'Fredoka One', cursive;
 
-	.__tab-wrap {
-		background-color: gray;
-	}
+	// :hover {
+	// 	background-color: darken($tab-color, $amount: 10);
+	// }
 }
 
 #chat-container {
@@ -23,7 +24,7 @@
 	
 	display: flex;
 	width: 100%;
-	height: 100%;
+	height: 100vh;
 	
 
 	.left-sidebar {
@@ -34,10 +35,11 @@
 		color: white;
 
 		.title {
-			font-size: large;
+			font-size: 20px;
 			justify-content: center;
 			align-items: center;
 			align-content: center;
+			border-bottom: black;
 
 			.add-icon {
 				cursor: pointer;
@@ -100,13 +102,13 @@
 				color: white;
 				padding: 0px 0px 0px 15px;
 				width: 56%;
-				height: 5%;
+				// height: 5%;
 				background: rgba($color: rgb(135, 135, 135), $alpha: 0.5);
 				border-radius: 10px;
 				border: none;
 				bottom: 0;
 				width: 98%;
-				height: 70px;
+				height: 50px;
 			}
 		}
 	
@@ -116,8 +118,6 @@
 		width: 20%;
 		padding: 10px;
 		background-color: rgba(32, 32, 38, 0.7);
-		
-
 	}
 
 button {
